backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
mysql-connector-python==9.1.0
opencv-python==4.11.0.86
numpy==2.2.1
mediapipe==0.10.21
requests==2.32.3
python-dotenv==1.0.1
